requirements.txt
langchain-core==0.3.9
langchain-community==0.3.1
google-cloud-spanner==3.49.1
langchain-experimental==0.3.2
